- Removing old docs that are not very useful - Converted a bunch of docs from APT to Markdown. Converting to HTML and then using Pandoc seems to work quite well.
End goal is to have a site that can be mirrored to Git and have all the pages render by just browser. I hope this will
encourage contribution to our site like Github is doing for the rest of our projects. I consider APT dead and it's just
another barrier to having people try and contribute documentation. Everyone knows Markdown and the tool support for it 
amazing. I made a mistake many years ago... I looked at Markdown and the perl script and said "no one is ever going to use that." :-)



git-svn-id: https://svn.apache.org/repos/asf/maven/site/trunk@1606945 13f79535-47bb-0310-9956-ffa450edef68