Adding new logo vector original source and png image versions.
This is an original composition created by me, Stephen Connolly.
The Apache Feather logo, which is a trademark of the Apache Software Foundation 
forms part of this work.

I am assigning the trademark rights of the image as a whole to The Apache Software Foundation
(Obviously the trademark rights of the feather logo are already held by The Apache Software
Foundation). The trademark rights of the composition are intended for use by the Apache
Maven project, and the Apache Maven PMC will be responsible for maintaining those rights.

Note: this composition is being committed under the terms of the Apache License, Version 2.0
The grant of trademark rights is strictly limited to the Apache Software Foundation.


git-svn-id: https://svn.apache.org/repos/asf/maven/site/trunk@1643777 13f79535-47bb-0310-9956-ffa450edef68