Adding mascot vector original and png images
This is an original work of art created by me, Stephen Connolly.

I am assigning the trademark rights of the image to The Apache Software Foundation
for use by the Apache Maven project, and the Apache Maven PMC will be responsible for maintaining those rights.

Note: this composition is being committed under the terms of the Apache License, Version 2.0

git-svn-id: https://svn.apache.org/repos/asf/maven/site/trunk@1650247 13f79535-47bb-0310-9956-ffa450edef68