document all allowed elements inside profiles
Previously the list was incomplete but did not infer the list was partial, e.g.
by using the word "some". If a user comes across the introductory guide first,
then it is easy to miss the complete model documentation and not find the
complete set of valid elements.

As discussed, the following have been updated:
- the remaining few elements have been added to complete the list.
- a link to the full model documentation is now provided.