Update guide-attached-tests.apt
The guide is incomplete.  Surefire does not pick up the tess in the attached jar on its own, it has to be configured to to so, and its provider autodetection does not work with attached jars, so it has to be nudged.

With this change, the guide would have saved me an hour.