Document the 'security model' for Maven
To make it easier for users to understand what to expect, and for
security researchers to decide where to focus their efforts.

I guess we could add further nuance when describing particular
security features (e.g. perhaps we don't have to trust repositories
themselves if we closely check the signatures on all downloaded
material), but this might be a good starting point for generally
setting the right expectations.