Set both maven.compiler.source and maven.compiler.target to 1.8 in maven-in-five-minutes.apt
Since Maven 3.9.6 does not support 1.7 compiler, the sample project cannot be built as is. This is frustrating for an inexperienced user. The proposed fix was tested on Java 21.0.3, Win7 sp1+.
P.S. Notice, the problem is not only with the docs. 
The code issue is known but still unresolved:
https://issues.apache.org/jira/projects/MARCHETYPES/issues/MARCHETYPES-82?filter=allopenissues
Even though the updated version was pushed to Maven Central Repository years ago:
https://github.com/ywchang/maven-archetype-quickstart
https://central.sonatype.com/artifact/com.github.ywchang/maven-archetype-quickstart
In a meantime, the new user gets a build failure trying Maven for their first time.