#MNGSITE-539 Add Maven CI Friendly Versions to menu
The guide about Maven CI Friend Versions was integrated into the navigation menu.
As it is a guide the file was moved to the guide section and renamed to match the naming schema of this folder.
Therefor the release notes of Maven 3.5.0 (and its beta) was updated to the new links.
Also some small formatting errors in those release notes were fixed in the same action.