MNGSITE-534 Mention Maven Daemon
This PR adds a new section "Maven Tools" to the site's navigation to create a place about tools aside Maven itself, similar to the plugin page.
Initially only the Maven Daemon is listed on the index page, but in the future
more tools can be added, e.g. Maven Shell, Maven Encrypt etc.