User guide: Migrate projects to Maven 4
PR includes a guide for users how to migrate projects to Maven 4.
Goal is to provide a guideline, which gets continuously updated in the future, when people share their experiences and issues.

As it's a user guide it was placed in the user section. Also the "What's new in Maven 4" article, which also aims at users as an audience, was moved there (in terms of navigation). To give both some visibility, when people open the Maven site, links were added to the landing page.

closes #729